requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
python-dotenv==1.0.0
flasgger==0.9.7.1
psycopg2-binary==2.9.9
flask-limiter==3.5.0
